Rename files to use snake-case
We use snake-case for modules in most of our repos, and it generally
seems more common for the npm ecosystem. This is a good time to change
this, now that we're changing the file locations already with the
TypeScript conversion.
