+++ src/validate-changelog.test.ts
@@ -1,4 +1,4 @@
-import { validateChangelog } from './validateChangelog';
+import { validateChangelog } from './validate-changelog';
 
 const emptyChangelog = `# Changelog
 All notable changes to this project will be documented in this file.
